refactor(scene3d): 抽出工具栏到 scene3dToolbar.tsx [split 1/N]
PanelButton/SceneAddButton/CanvasPanelRestoreButton/SceneAddToolbar 整体移出,
逻辑逐字不变;壳 3822→3478。双 tsc 零错误,scene3d 测试绿
